% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export Highcharts widget to PNG}
\usage{
export(
  hc,
  width = NULL,
  height = NULL,
  page_width = 8.27,
  page_height = 11.69,
  factor_width = 1,
  factor_height = 1/2,
  dpi = 96,
  filename = "plot.png",
  delay = 0.5
)
}
\arguments{
\item{hc}{htmlwidget object}

\item{width}{numeric}

\item{height}{numeric}

\item{page_width}{numeric default 8.27 inches (A4)}

\item{page_height}{numeric default 11.69 inches (A4)}

\item{factor_width}{numeric default 1. Full width}

\item{factor_height}{numeric default 1/2. Half page}

\item{dpi}{numeric default 96}

\item{filename}{character default "plot.png"}

\item{delay}{numeric default 0.5}
}
\description{
Export Highcharts widget to PNG
}
