% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_chk_coerce_date.R
\name{ngr_chk_coerce_date}
\alias{ngr_chk_coerce_date}
\title{Check and Validate Date-Coercible Input}
\usage{
ngr_chk_coerce_date(x, x_name = NULL)
}
\arguments{
\item{x}{\link{any} An object expected to be a \link{Date} or a string coercible to a
\link{Date}.}

\item{x_name}{\link{character} Optional. A single string giving the name of \code{x}
used in error messages. If \code{NULL}, the name is inferred from the calling
expression.}
}
\value{
\code{x} invisibly if it is a \link{Date} or coercible to one. An error is thrown if
coercion fails.
}
\description{
Checks whether an object is a \link{Date} or can be safely coerced to a \link{Date}
using \code{\link[base:as.Date]{base::as.Date()}}. If the input is coercible, it is returned
invisibly; otherwise, an informative error is thrown.
}
\details{
This helper is intended for lightweight validation in functions that accept
date inputs but do not need to modify them. Coercion is attempted via
\code{\link[base:as.Date]{base::as.Date()}}, and failure results in a call to
\code{\link[chk:abort_chk]{chk::abort_chk()}}.
}
\seealso{
\code{\link[base:as.Date]{base::as.Date()}}, \code{\link[chk:abort_chk]{chk::abort_chk()}}

Other chk: 
\code{\link{ngr_chk_dt_complete}()}
}
\concept{chk}
