% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_dbqs_ltree.R
\name{ngr_dbqs_ltree}
\alias{ngr_dbqs_ltree}
\title{Cast Character to PostgreSQL ltree}
\usage{
ngr_dbqs_ltree(x)
}
\arguments{
\item{x}{\link{character} A single string or character vector to be cast to \code{ltree}.}
}
\value{
\link{character} A string or character vector of PostgreSQL ltree-compatible text. Each element is wrapped in single quotes and cast with \verb{::ltree}.
}
\description{
Constructs a PostgreSQL ltree literal from a character string.
This is useful for composing queries that rely on ltree path syntax.
}
\details{
This function is typically used to help construct PostgreSQL queries that operate on \code{ltree} data types.
It simply formats each input string into the PostgreSQL \code{ltree} cast syntax.
}
\examples{
ngr_dbqs_ltree("400.431358.623573")

}
\concept{Database Query}
