% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_dbqs_tbl_quote.R
\name{ngr_dbqs_tbl_quote}
\alias{ngr_dbqs_tbl_quote}
\title{Quote a Table Name for SQL Queries}
\usage{
ngr_dbqs_tbl_quote(table_name, quote_type = "double")
}
\arguments{
\item{table_name}{\link{character} A single string representing the table name to be quoted.}

\item{quote_type}{\link{character} A single string specifying the type of quote to use, either \code{'single'} or \code{'double'}. Default is \code{'double'}.}
}
\value{
\link{character} The table name wrapped in the specified type of quotation marks.
}
\description{
This function wraps a table name with the specified type of quotation marks, ensuring compatibility with SQL queries.
}
\details{
The function uses \code{glue::glue()} to format the table name with either single or double quotes. If an invalid \code{quote_type} is specified, an error is thrown.
}
\examples{
ngr_dbqs_tbl_quote("my_table")
ngr_dbqs_tbl_quote("my_table", quote_type = "single")

}
