% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_git_issue_details.R
\name{ngr_git_issue_details}
\alias{ngr_git_issue_details}
\title{Get GitHub Issue Timeline or Event Details}
\usage{
ngr_git_issue_details(
  issue_url_json,
  token = Sys.getenv("GITHUB_PAT"),
  endpoint = "timeline",
  events_all = FALSE
)
}
\arguments{
\item{issue_url_json}{\link{character}. The full GitHub API URL of the issue (e.g., "https://api.github.com/repos/user/repo/issues/123").}

\item{token}{\link{character}. A GitHub personal access token (PAT). Default is \code{Sys.getenv("GITHUB_PAT")}.}

\item{endpoint}{\link{character}. The API endpoint to use for retrieving issue events. Defaults to "timeline".
to access administrative actions like mentioned, labeling or assigning.}

\item{events_all}{\link{logical}. If \code{TRUE}, return all events. If \code{FALSE} (default), return only those associated with commits (i.e., with a non-null \code{commit_id}).}
}
\value{
A tibble with columns: \code{issue_url_json}, \code{event}, \code{commit_id}, \code{commit_url_json}, and \code{created_at}.
}
\description{
This function retrieves the timeline or event details for a specified GitHub issue using the GitHub API. By default, it filters for events associated with commits only.
}
\examples{
ngr_git_issue_details("https://api.github.com/repos/NewGraphEnvironment/fpr/issues/13")
ngr_git_issue_details("https://api.github.com/repos/NewGraphEnvironment/fpr/issues/13", events_all = TRUE)
}
\references{
GitHub Issue Event Types: https://docs.github.com/en/rest/using-the-rest-api/issue-event-types
}
\seealso{
\code{\link[gh:gh]{gh::gh()}}, \code{\link[purrr:map_dfr]{purrr::map_dfr()}}, \code{\link[tibble:tibble]{tibble::tibble()}}
}
