% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_hyd_realtime.R
\name{ngr_hyd_realtime}
\alias{ngr_hyd_realtime}
\title{Retrieve Realtime HYDAT Data for a Station}
\usage{
ngr_hyd_realtime(
  id_station = NULL,
  param_primary = c(5),
  param_secondary = c(1, 18, 19, 41, 46, 6, 7, 8, 10, 40, 47),
  days_back = 581
)
}
\arguments{
\item{id_station}{\link{character} A single station identifier string.}

\item{param_primary}{\link{numeric} A numeric vector of primary HYDAT parameter codes.}

\item{param_secondary}{\link{numeric} A numeric vector of secondary HYDAT parameter codes. This data is only retrieved if
there is primary data available. See the available options with \link[tidyhydat:param_id]{tidyhydat::param_id}}

\item{days_back}{\link{numeric} A single integer specifying how many days back to fetch data. Defaults to 581 which seems
to be the maximum amount retrievable}
}
\value{
A list with two elements: \code{primary_data} and \code{secondary_data}, each a data frame of results. Returns \code{NULL} on failure and prints a CLI alert.
}
\description{
Convenience function to fetch primary and secondary water survey parameters for a given HYDAT station
over a specified number of days. Wraps \code{\link[tidyhydat:realtime_ws]{tidyhydat::realtime_ws()}} with main advantage that it relies on the
availability of temperature data (or some other \code{param_primary} to be present before it will attempt to obtain
secondary data.
}
\seealso{
\code{\link[tidyhydat:realtime_ws]{tidyhydat::realtime_ws()}}
}
