% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_pkg_detach.R
\name{ngr_pkg_detach}
\alias{ngr_pkg_detach}
\title{Detach a package}
\usage{
ngr_pkg_detach(pkg)
}
\arguments{
\item{pkg}{\link{character} A single string representing the name of the package to detach.}
}
\value{
Invisibly returns NULL.
}
\description{
This function is a convenience wrapper around \verb{[detach()]} to unload a package.
It is designed to speed up development and testing for package development and report writing.
}
\examples{
\dontrun{
ngr_pkg_detach("ngr")
}
}
\concept{package}
