% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_s3_files_to_index.R
\name{ngr_s3_files_to_index}
\alias{ngr_s3_files_to_index}
\title{Generate S3 Index HTML}
\usage{
ngr_s3_files_to_index(
  files,
  dir_output = ".",
  filename_output = "index.html",
  ask = TRUE,
  header1 = "Index"
)
}
\arguments{
\item{files}{\link{character} A character vector of file and directory names (e.g., \code{"file1.txt"}, \code{"folder/"}).}

\item{dir_output}{\link{character} A string specifying the output directory where the file will be created. Defaults to the current working directory (\code{"."}).}

\item{filename_output}{\link{character} A string specifying the name of the output file. Defaults to \code{"index.html"}.}

\item{ask}{\link{logical} A logical value indicating whether to prompt the user before overwriting an existing file. Defaults to \code{TRUE}.}

\item{header1}{\link{character} A string specifying the \verb{<h1>} title for the HTML page. Defaults to \code{"Index"}.}
}
\value{
\link{character} Invisibly returns the path to the generated file.
}
\description{
This function generates an \code{index.html} file to link to an S3 bucket's contents. Can be loaded to bucket to provide
a simple web interface for browsing files. For now we need to be sure to run this in the cmd line to see it though
\verb{aws s3 website s3://\{my_bucket_name\}/ --index-document index.html}
}
\details{
This function uses \code{\link[=ngr_s3_path_to_https]{ngr_s3_path_to_https()}} to convert file paths into HTTPS links before constructing the HTML.
}
\examples{
# Example input
files <- c("file1.txt", "file2.txt", "folder1/", "folder2/")

# Specify a custom output directory and filename
ngr_s3_files_to_index(files, dir_output = "output_dir", filename_output = "custom_index.html", header1 = "My Custom Index")
}
\concept{S3}
