% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_s3_path_to_https.R
\name{ngr_s3_path_to_https}
\alias{ngr_s3_path_to_https}
\title{Convert an S3 Path to an https URL}
\usage{
ngr_s3_path_to_https(s3_path, website = FALSE, aws_region = "us-west-2")
}
\arguments{
\item{s3_path}{\link{character} A single S3 path to convert, starting with \code{"s3://"}.}

\item{website}{\link{logical} Whether to construct a static website hosting URL. Default is \code{FALSE}. This can be used to
serve an index.html file via http (unsecured) like the ones created with \code{\link[=ngr_s3_files_to_index]{ngr_s3_files_to_index()}}.}

\item{aws_region}{\link{character} The AWS region for the bucket. Default is \code{"us-west-2"}.}
}
\value{
\link{character} The https or http URL corresponding to the S3 path.
}
\description{
This function converts an S3 path into its corresponding https or website-specific http URL.
}
\details{
The function removes the \code{"s3://"} prefix, splits the remaining path into the bucket name and object key, and
constructs an https URL in the format \verb{https://<bucket>.s3.amazonaws.com/<key>}. If \code{website = TRUE}, it constructs
a http URL in the format \verb{http://<bucket>.s3-website-<region>.amazonaws.com/<key>}.

Note: HTTPS is not supported for static website URLs directly from S3; use CloudFront for HTTPS.
}
\examples{
# Example usage:
ngr_s3_path_to_https("s3://my-bucket/my-object.txt")
ngr_s3_path_to_https("s3://my-bucket/my-object.txt", website = TRUE)
}
\concept{s3}
