% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_df_detect_filter.R
\name{ngr_str_df_detect_filter}
\alias{ngr_str_df_detect_filter}
\title{Filter Rows Based on String Detection in a Column Detected with a partial String Match}
\usage{
ngr_str_df_detect_filter(x, col_filter, str_filter)
}
\arguments{
\item{x}{\link{data.frame} The input data frame to filter.}

\item{col_filter}{\link{character} A single string representing the pattern to match column names.}

\item{str_filter}{\link{character} A single string to match within the values of the filtered column.}
}
\value{
A filtered data frame if matching rows are found; otherwise, the original data frame is returned.
}
\description{
This function filters rows in a data frame where a specified column's name matches a pattern and the values of that column contain a specified string.
}
\seealso{
Other string: 
\code{\link{ngr_str_dir_from_path}()},
\code{\link{ngr_str_link_repo}()}
}
\concept{string}
