% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_dir_from_file.R
\name{ngr_str_dir_from_file}
\alias{ngr_str_dir_from_file}
\title{Detect the Presence of a Specific File in a Directory}
\usage{
ngr_str_dir_from_file(path_dir, file_name, recurse = FALSE, ...)
}
\arguments{
\item{path_dir}{\link{character} A single string representing the directory path to search.}

\item{file_name}{\link{character} A string specifying the name of the file to detect.}

\item{recurse}{\link{logical} Whether to search recursively in subdirectories. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[fs:dir_ls]{fs::dir_ls()}}.}
}
\value{
\link{character} The directory path if the file is found; otherwise, a warning is issued.
}
\description{
This function checks if a specific file exists in a given directory and optionally its subdirectories.
}
