% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_dir_from_path.R
\name{ngr_str_dir_from_path}
\alias{ngr_str_dir_from_path}
\title{Extract Directory Name from File Path Based on Levels}
\usage{
ngr_str_dir_from_path(path, levels = 2)
}
\arguments{
\item{path}{\link{character} A single string. The file path from which to extract the directory name.}

\item{levels}{\link{integer} A single integer. The number of levels up to traverse from the file path to extract the directory name. Default is \code{2}.}
}
\value{
\link{character} A single string representing the directory name at the specified level.
}
\description{
This function retrieves the directory name at a specified level(s) above the given file path.
}
\examples{
ngr_str_dir_from_path("~/Projects/repo/ngr/data-raw/extdata.R", levels = 2)
# Returns "ngr"
}
\seealso{
Other string: 
\code{\link{ngr_str_df_detect_filter}()},
\code{\link{ngr_str_repo_link}()}
}
\concept{string}
