% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_link_repo.R
\name{ngr_str_link_repo}
\alias{ngr_str_link_repo}
\title{Create an HTML Link to Repository Resources or GitHub Pages}
\usage{
ngr_str_link_repo(
  repo_source,
  url = "https://www.newgraphenvironment.com",
  anchor_text = "url_link",
  target = "_blank"
)
}
\arguments{
\item{repo_source}{\link{character} A single string representing the repository name to be linked.}

\item{url}{\link{character} A single string specifying the base URL for the repository host. Default is New Graph
gitpages at \url{https://www.newgraphenvironment.com"}. \url{https://github.com/NewGraphEnvironment} gets you to the
repository itself.}

\item{anchor_text}{\link{character} A single string specifying the text displayed for the link. Defaults to \code{"url_link"}.}

\item{target}{\link{character} A single string indicating the \code{target} attribute in the HTML link. Default is \code{"_blank"}. Options include:
\itemize{
\item \code{"_blank"}: Opens the link in a new tab or window.
\item \code{"_self"}: Opens the link in the same tab or window.
\item \code{"_parent"}: Opens the link in the parent frame, if applicable.
\item \code{"_top"}: Opens the link in the full body of the window, ignoring frames.
}}
}
\value{
\link{character} A single string containing the HTML \verb{<a>} tag.
}
\description{
This function generates an HTML \verb{<a>} tag linking to a specified repository resource or GitHub Pages site hosted from the repository.
It is particularly useful for embedding links in HTML documents.
}
\details{
This function can create links to either:
\itemize{
\item \strong{Repository resources}: Links directly to the repository on platforms like GitHub.
\item \strong{GitHub Pages}: Links to websites or documentation served from the repository, using a base URL for the hosted site.
}

If the provided \code{url} does not start with \verb{http://} or \verb{https://}, a warning is issued
to indicate potential issues with the generated links.
}
\examples{
# Example 1: Link to the repository with default anchor text
ngr_str_link_repo("ngr", url = "https://github.com/NewGraphEnvironment")

# Example 2: Link to GitHub Pages with anchor text set to the repository name
ngr_str_link_repo("ngr", url = "https://www.newgraphenvironment.com", anchor_text = "ngr")

}
\seealso{
Other string: 
\code{\link{ngr_str_df_detect_filter}()},
\code{\link{ngr_str_dir_from_path}()}
}
\concept{string}
