% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_link_url.R
\name{ngr_str_link_url}
\alias{ngr_str_link_url}
\title{Create an HTML Link to Repository Resources or GitHub Pages}
\usage{
ngr_str_link_url(
  repo_source = NULL,
  url = "https://www.newgraphenvironment.com",
  anchor_text = "url_link",
  target = "_blank"
)
}
\arguments{
\item{repo_source}{\link{character} A single string representing the repository name to be linked. Optional. Default is \code{NULL}.}

\item{url}{\link{character} A single string specifying the either the URL to link to or (if \code{repo_source} is provided)
the base URL for the repository host. Default is New Graph
gitpages at \url{https://www.newgraphenvironment.com"}. \url{https://github.com/NewGraphEnvironment} gets you to the
repository itself.}

\item{anchor_text}{\link{character} A single string specifying the text displayed for the link. Defaults to \code{"url_link"}.}

\item{target}{\link{character} A single string indicating the \code{target} attribute in the HTML link. Default is \code{"_blank"}. Options include:
\itemize{
\item \code{"_blank"}: Opens the link in a new tab or window.
\item \code{"_self"}: Opens the link in the same tab or window.
\item \code{"_parent"}: Opens the link in the parent frame, if applicable.
\item \code{"_top"}: Opens the link in the full body of the window, ignoring frames.
}}
}
\value{
\link{character} A single string containing the HTML \verb{<a>} tag.
}
\description{
This function generates an HTML \verb{<a>} tag linking to a specified repository resource or GitHub Pages site hosted from the repository.
It is particularly useful for embedding links in HTML documents.
}
\examples{
# Example 1: Link to the repository with default anchor text
ngr_str_link_url("ngr", url = "https://github.com/NewGraphEnvironment")

# Example 2: Link to GitHub Pages with anchor text set to the repository name
ngr_str_link_url(repo_source = "ngr", anchor_text = "ngr")

# Example 3: Link with no repo_source
ngr_str_link_url(url = "https://www.newgraphenvironment.com", anchor_text = "Visit New Graph")

}
\seealso{
Other string: 
\code{\link{ngr_str_df_detect_filter}()},
\code{\link{ngr_str_dir_from_path}()}
}
\concept{string}
