% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_str_replace_filenames.R
\name{ngr_str_replace_filenames}
\alias{ngr_str_replace_filenames}
\title{Rename Files and Directories Containing Specific Text}
\usage{
ngr_str_replace_filenames(
  path = ".",
  text_current,
  text_replace,
  glob_exclude = ".git",
  glob_include = NULL,
  ask = TRUE
)
}
\arguments{
\item{path}{\link{character} A vector of one or more paths. The root directory to start searching for files and directories.}

\item{text_current}{\link{character} A string representing the text to be replaced in filenames.}

\item{text_replace}{\link{character} A string representing the new text to replace the current text in filenames.}

\item{glob_exclude}{\link{character} A glob pattern to exclude specific paths from renaming. Default is ".git".}

\item{glob_include}{\link{character or NULL} A glob pattern to include specific files or directories. Default is \code{NULL}.}

\item{ask}{\link{logical} Whether to prompt the user for confirmation before renaming. Default is \code{TRUE}.}
}
\value{
\code{NULL} invisibly after completing the renaming operation.
}
\description{
This function renames files and directories by replacing occurrences of a specified text with a new text.
It allows exclusions and inclusions using glob patterns and provides an option to confirm changes before applying them.
}
\details{
The function uses \code{stringr::str_detect()} to find filenames containing the specified \code{text_current} and excludes those where \code{text_replace} is already present. It then renames files and directories accordingly using \code{fs::file_move()}.
}
\concept{replace}
