% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_tidy_cols_rm_na.R
\name{ngr_tidy_cols_rm_na}
\alias{ngr_tidy_cols_rm_na}
\title{Remove Columns with All NA Values}
\usage{
ngr_tidy_cols_rm_na(df)
}
\arguments{
\item{df}{A \link{data.frame}. The input data frame to be cleaned.}
}
\value{
A \link{data.frame} with columns containing only \code{NA} values removed.
}
\description{
\code{ngr_tidy_cols_rm_na} removes all columns from a \link{data.frame} where every value is \code{NA}.
}
\examples{
# Example data frame
df <- data.frame(
  A = c(NA, NA, NA),
  B = c(1, 2, NA),
  C = c(NA, NA, NA),
  D = c("x", "y", NA)
)

# Apply the function
tidy_df <- ngr_tidy_cols_rm_na(df)

# Result
print(tidy_df)
}
\concept{tidy}
