% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_tidy_type.R
\name{ngr_tidy_type}
\alias{ngr_tidy_type}
\title{Coerce column types in one data frame to match another}
\usage{
ngr_tidy_type(dat_w_types, dat_to_type)
}
\arguments{
\item{dat_w_types}{\link{data.frame} Reference data frame with desired column types.}

\item{dat_to_type}{\link{data.frame} Data frame to be coerced to match \code{dat_w_types}.}
}
\value{
\link{data.frame} \code{dat_to_type} with shared columns coerced to the types in \code{dat_w_types}.
}
\description{
This function ensures that shared columns in \code{dat_to_type} are coerced to match
the data types of corresponding columns in \code{dat_w_types}.
}
\examples{
dat_w_types <- data.frame(a = as.numeric(1:3), b = as.character(4:6))
dat_to_type <- data.frame(a = as.character(1:3), b = 4:6)
ngr_tidy_type(dat_w_types, dat_to_type)
}
\seealso{
Other tidy: 
\code{\link{ngr_tidy_cols_rm_na}()},
\code{\link{ngr_tidy_cols_type_compare}()}
}
\concept{tidy}
