% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_xl_map_colnames.R
\name{ngr_xl_map_colnames}
\alias{ngr_xl_map_colnames}
\title{Extract Formulas from an Excel Worksheet}
\usage{
ngr_xl_map_colnames(path, sheet, row_id)
}
\arguments{
\item{path}{\link{character} A path to the Excel file to be read. Must be a valid file path.}

\item{sheet}{\link{character} The name of the sheet from which to extract formulas. Must be specified.}

\item{row_id}{\link{numeric} The row number to filter names by. Must be specified.}
}
\value{
A data frame of column names, column names cleaned (lower snakecase) their associated excel column id (address),
excel address without a row number
(for tying to formulas).
}
\description{
This function reads rownames from a specific sheet and row of an Excel file, maps them to the excel column id (address).
}
\details{
It uses \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}} to parse the Excel file and filters cells based on the provided sheet name and row ID.
Removes row ids from excel addresses and tidies column names using \code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}

This function reads the Excel file and filters cells to extract only those with formulas
from the specified sheet and row. The output includes all unique formulas from the target row.
}
\examples{
path <- system.file("extdata", "pscis_phase1.xlsm", package = "ngr")
ngr_xl_map_colnames(path, sheet = "PSCIS Assessment Worksheet", row_id = 4)

}
\seealso{
Other excel: 
\code{\link{ngr_xl_map_formulas}()},
\code{\link{ngr_xl_read_formulas}()}
}
\concept{excel}
