% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_xl_map_formulas.R
\name{ngr_xl_map_formulas}
\alias{ngr_xl_map_formulas}
\title{Read Formulas and Column Names from an Excel Worksheet and Combine to Aid in Transformation of Formulas to R}
\usage{
ngr_xl_map_formulas(path, sheet, rowid_colnames, rowid_formulas)
}
\arguments{
\item{path}{\link{character} A path to the Excel file to be read. Must be a valid file path.}

\item{sheet}{\link{character} The name of the sheet from which to extract formulas and/or row names. Must be specified.}

\item{rowid_colnames}{\link{numeric} A single row number used to extract column names.}

\item{rowid_formulas}{\link{numeric} A vector of row numbers to filter formulas by. Must be specified.}
}
\value{
A data frame with the following columns:
\describe{
\item{address_rowless}{Excel address without the row ID appended.}
\item{address}{Raw Excel column identifier.}
\item{row}{Row number in the Excel spreadsheet from which the formula was extracted.}
\item{formula}{Raw formula as it appears in the Excel spreadsheet.}
\item{formula_rowless}{Formula with row numbers of referenced cells removed, aligned with \code{address_rowless}.}
\item{formula_colname}{Column name indicating where the formula is located in the Excel spreadsheet.}
\item{formula_with_col_names}{Excel formula with associated column names substituted for Excel addresses.}
}
}
\description{
This function reads single row formulas (formulas working on columns within the same row of the spreadsheet where the
fomula is located) and column names from a specific sheet and rows of an Excel file.
It uses \code{tidyxl::xlsx_cells()} to parse the Excel file then uses \code{ngr} helper functions \code{\link[=ngr_xl_read_formulas]{ngr_xl_read_formulas()}}
and \code{\link[=ngr_xl_map_colnames]{ngr_xl_map_colnames()}} to clean up formulas and column names. Column names are subbed in the place of
excel addresses so (for now) the formulas can be passed to LLMs to convert from excel syntax to \code{R}.
}
\examples{
path <- system.file("extdata", "pscis_phase1.xlsm", package = "ngr")
result <- ngr_xl_map_formulas(path, sheet = "PSCIS Assessment Worksheet", rowid_colnames = 4, rowid_formulas = c(5, 6))
result[, c("formula_colname", "formula_with_col_names")][7, ]

}
\seealso{
Other excel: 
\code{\link{ngr_xl_map_colnames}()},
\code{\link{ngr_xl_read_formulas}()}
}
\concept{excel}
