% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_xl_read_formulas.R
\name{ngr_xl_read_formulas}
\alias{ngr_xl_read_formulas}
\title{Extract Single Row Formulas from an Excel Worksheet}
\usage{
ngr_xl_read_formulas(path, sheet, row_id)
}
\arguments{
\item{path}{\link{character} A path to the Excel file to be read. Must be a valid file path.}

\item{sheet}{\link{character} The name of the sheet from which to extract formulas. Must be specified.}

\item{row_id}{\link{numeric} The row number to filter formulas by. Must be specified. Can be a vector of rows (ie. c(5,6))}
}
\value{
A data frame of unique formulas with their associated metadata for the specified sheet and row.
}
\description{
This function reads single row formulas (formulas working on columns within the same row of the spreadsheet where the
fomula is located) from a specific sheet and row of an Excel file.
It uses \code{\link[tidyxl:xlsx_cells]{tidyxl::xlsx_cells()}} to parse the Excel file and filters cells based on the provided sheet name and row ID.
It processes the formulas to remove row specific addresses so that they can be abstracted to R with less friction.
Designed as a helper function to work with \code{\link[=ngr_xl_map_colnames]{ngr_xl_map_colnames()}}.
}
\details{
This function reads the Excel file and filters cells to extract only those with formulas
from the specified sheet and row. The output includes all unique formulas from the target row.
}
\examples{
path <- system.file("extdata", "pscis_phase1.xlsm", package = "ngr")
res <- ngr_xl_read_formulas(path, sheet = "PSCIS Assessment Worksheet", row_id = c(5,6))
res[, "formula_rowless"]

}
\seealso{
Other excel: 
\code{\link{ngr_xl_map_colnames}()},
\code{\link{ngr_xl_map_formulas}()}
}
\concept{excel}
