% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showHTML.R
\name{showHTML}
\alias{showHTML}
\title{Display HTML content in iframe}
\usage{
showHTML(
  url,
  height = "600px",
  width = "100\%",
  show_link = TRUE,
  link_text = "URL"
)
}
\arguments{
\item{url}{URL of the HTML content to display}

\item{height}{Height of iframe (default "600px")}

\item{width}{Width of iframe (default "100\%")}

\item{show_link}{Show link to open in new window (default TRUE)}

\item{link_text}{Custom link text (default "URL")}
}
\value{
Prints iframe HTML using cat() for results: asis
}
\description{
Display HTML content in iframe
}
