% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr-package.R
\docType{package}
\name{ngr-package}
\alias{ngr}
\alias{ngr-package}
\title{ngr: New Graph Reporting}
\description{
Functions used in dynamic reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/ngr}
  \item \url{https://newgraphenvironment.github.io/ngr/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/ngr/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

}
\keyword{internal}
