% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_s3_keys_get.R
\name{ngr_s3_keys_get}
\alias{ngr_s3_keys_get}
\title{List S3 Keys from a Public or Authenticated S3 Bucket}
\usage{
ngr_s3_keys_get(url_bucket, prefix = NULL, pattern = NULL)
}
\arguments{
\item{url_bucket}{\link{character} A single string. The URL of the S3 bucket (e.g., "https://nrs.objectstore.gov.bc.ca/gdwuts").}

\item{prefix}{\link{character} Optional. A prefix to filter the S3 keys returned. Default is \code{NULL}.}

\item{pattern}{\link{character} Optional. A vector of glob or substring patterns used to filter the result. If patterns contain wildcards (\code{*} or \verb{?}), they are treated as glob; otherwise as fixed substring matches.}
}
\value{
\link{character} A character vector of fully qualified S3 object URLs that match the criteria. If no matches are found or the bucket fails to respond, an error is raised or an empty character vector is returned.
}
\description{
Retrieve all object keys from an Amazon S3 bucket, optionally filtered by prefix and glob or substring pattern.
}
\details{
This function paginates through the XML responses from S3 using the \code{marker} parameter until \code{IsTruncated} is \code{FALSE}. Keys can be filtered using both prefix and flexible glob or fixed-string matching. Final returned values are constructed by joining \code{url_bucket} with each key.
}
\examples{
keys <- ngr_s3_keys_get(
  url_bucket = "https://nrs.objectstore.gov.bc.ca/gdwuts",
  prefix = "093/093l",
  pattern = c("dem", "*.tif")
)
head(keys)
}
\seealso{
\code{\link[httr2:request]{httr2::request()}}, \code{\link[xml2:read_xml]{xml2::read_xml()}}, \code{\link[utils:glob2rx]{utils::glob2rx()}}
}
