% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngr_spk_rast_not_empty.R
\name{ngr_spk_rast_not_empty}
\alias{ngr_spk_rast_not_empty}
\title{Check if Raster Has Non-Zero Data}
\usage{
ngr_spk_rast_not_empty(f)
}
\arguments{
\item{f}{\link{character} A single file path to a raster file.}
}
\value{
\link{logical} \code{TRUE} if the raster has any non-zero values, otherwise \code{FALSE}.
}
\description{
Determines whether a raster file contains any non-zero values.
}
\seealso{
Other spacehakr: 
\code{\link{ngr_spk_gdalwarp}()},
\code{\link{ngr_spk_join}()},
\code{\link{ngr_spk_layer_info}()},
\code{\link{ngr_spk_odm}()},
\code{\link{ngr_spk_poly_to_points}()},
\code{\link{ngr_spk_q_layer_info}()},
\code{\link{ngr_spk_rast_ext}()},
\code{\link{ngr_spk_rast_rm_empty}()},
\code{\link{ngr_spk_res}()},
\code{\link{ngr_spk_stac_calc}()}
}
\concept{spacehakr}
